Increment1 Add test for GET handler
