+++ go.mod
@@ -2,4 +2,7 @@ module github.com/sashganush/shortcut
 
 go 1.17
 
-require github.com/go-chi/chi/v5 v5.0.7
+require (
+	github.com/go-chi/chi/v5 v5.0.7
+	github.com/stretchr/testify v1.7.0
+)
