+++ internal/app/shortener/handlers_test.go
@@ -1,0 +1,47 @@
+package internal
+
+import (
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+)
+
+func TestUserViewHandler(t *testing.T) {
+	type want struct {
+		code        int
+		response    string
+		contentType string
+	}
+	tests := []struct {
+		name string
+		want want
+	}{
+		{
+			name: "positive test #1",
+			want: want{
+				code:        307,
+				response:    `https://google.com/abcd`,
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			request := httptest.NewRequest(http.MethodGet, "/4444", nil)
+			w := httptest.NewRecorder()
+			h := http.HandlerFunc(GetRequestHandler)
+			h.ServeHTTP(w, request)
+			result := w.Result()
+
+
+			assert.Equal(t, tt.want.code, result.StatusCode)
+
+			_, err := ioutil.ReadAll(result.Body)
+			require.NoError(t, err)
+			err = result.Body.Close()
+			require.NoError(t, err)
+		})
+	}
+}
