+++ internal/app/shortener/storage.go
@@ -1,0 +1,13 @@
+package internal
+
+const SHORTURLLEN = 10
+
+var allRedirects = map[string]string{
+	"4444": "https://google.com/abcd",
+	"1111": "https://ya.ru/1",
+	"kyQFSSqy": "http://t9xiost0mawj62.ru",
+	"ytOkkmjo": "http://y4qcfp8ur.yandex",
+
+}
+
+var letterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
